criei a validação para evitar que um produto que pertence a um pedido seja excluído.

+++ src/controllers/productController.js
@@ -5,6 +5,7 @@ const {
     deleteRegisterProduct,
     findProductsByCategoryId
 } = require("../database/productDatabase");
+const validatesWhetherTheProductBelongsToAnOrder = require("../services/deleteProductValidation");
 
 const registerProduct = async (req, res) => {
     const { descricao, quantidade_estoque, valor, categoria_id } = req.body;
@@ -80,9 +81,14 @@ const deleteProduct = async (req, res) => {
             return res.status(400).json({ mensagem: "O produto informado não existe." });
         };
 
+        const hasOrders = await validatesWhetherTheProductBelongsToAnOrder(productExist);
+        if (hasOrders) return res.status(403).json({ mensagem: "O produto informado está presente em um pedido e não pode ser excluído." });
+
         await deleteRegisterProduct(id);
         return res.status(200).json({ mensagem: "Produto excluido com sucesso." });
-    } catch (error) {
+    }
+
+    catch (error) {
         return res.status(500).json({ mensagem: error.message });
     }
 };
